back-end/pyproject.toml
[tool.poetry]
name = "chat-backend"
version = "0.1.0"
description = ""
authors = ["et-codes <104942211+et-codes@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aniso8601 = "9.0.1"
asgiref = "3.5.2"
bcrypt = "4.0.1"
bidict = "0.22.0"
certifi = "2022.9.24"
charset-normalizer = "2.1.1"
click = "8.1.3"
colorama = "0.4.5"
dnspython = "2.2.1"
eventlet = "0.33.1"
flask = "2.2.2"
flask-cors = "3.0.10"
flask-socketio = "5.3.1"
greenlet = "1.1.3.post0"
idna = "3.4"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.1"
psycopg2 = "2.9.4"
pyjwt = "2.6.0"
python-dotenv = "0.21.0"
python-engineio = "4.3.4"
python-socketio = "5.7.1"
pytz = "2022.4"
requests = "2.28.1"
six = "1.16.0"
urllib3 = "1.26.12"
werkzeug = "2.2.2"
gunicorn = {url = "https://github.com/benoitc/gunicorn/archive/refs/heads/master.zip#egg=gunicorn==20.1.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
